SCH-40 refactor file structure for database functions
